Restructured the project and added requirements.txt for the venv.
