--- cogs/__init__.py
@@ -1,2 +1,0 @@
-from cogs.ping import Ping
-from cogs.settings import Settings
