+++ docs/reference/nasa_api/nasa_api_errors.md
@@ -4,4 +4,4 @@ This is the file that stores all the classes that represent exceptions.
 
 ## NasaApiDataNotFoundError
 
-::: nasa_api.NasaApiDataNotFoundError
+::: src.nasa_api.NasaApiDataNotFoundError
