+++ src/cogs/settings.py
@@ -4,7 +4,7 @@ import discord
 from discord import app_commands
 from discord.ext import commands
 
-from utils.config import Config
+from src.utils.config import Config
 
 
 class Settings(commands.GroupCog, name="setting"):
